+++ packages/native-stack/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.0.2](https://github.com/react-navigation/react-navigation/compare/@react-navigation/native-stack@7.0.1...@react-navigation/native-stack@7.0.2) (2024-11-14)
+
+**Note:** Version bump only for package @react-navigation/native-stack
+
+## [7.0.1](https://github.com/react-navigation/react-navigation/compare/@react-navigation/native-stack@7.0.0...@react-navigation/native-stack@7.0.1) (2024-11-13)
+
+**Note:** Version bump only for package @react-navigation/native-stack
+
 # [7.0.0](https://github.com/react-navigation/react-navigation/compare/@react-navigation/native-stack@7.0.0-rc.30...@react-navigation/native-stack@7.0.0) (2024-11-06)
 
 **Note:** Version bump only for package @react-navigation/native-stack